+++ react-foldable/src/Foldable.tsx
@@ -10,30 +10,32 @@ const Foldable = (props: { children: React.ReactNode }) => {
   const isDualScreen = useIsDualScreen();
   const screenSpanning = useScreenSpanning();
 
-  return (
-    <FoldableContext.Provider
-      value={{ windowSegments, isDualScreen, screenSpanning }}
-    >
-      {React.Children.map(props.children, (c) => {
-        if (React.isValidElement<FoldableScreenProps>(c)) {
-          const { match, matchScreen } = c.props;
+  const contextValue = { windowSegments, isDualScreen, screenSpanning };
 
-          if (match) {
-            return match({ windowSegments, isDualScreen, screenSpanning })
-              ? React.cloneElement(c, {
-                  matchScreen: c.props.matchScreen,
-                })
-              : null;
-          } else if (matchScreen !== undefined) {
-            return matchScreen < windowSegments.length
-              ? React.cloneElement(c, {
-                  matchScreen: c.props.matchScreen,
-                })
-              : null;
-          } else {
-            return null;
+  return (
+    <FoldableContext.Provider value={contextValue}>
+      {React.Children.map(props.children, (foldableScreenChild) => {
+        if (React.isValidElement<FoldableScreenProps>(foldableScreenChild)) {
+          const { match, matchScreen } = foldableScreenChild.props;
+
+          // do we have a custom matcher function and does it match?
+          if (match && match(contextValue)) {
+            return React.cloneElement(foldableScreenChild, {
+              matchScreen: foldableScreenChild.props.matchScreen,
+            });
+            // do we have a screen to match and is it in the visible window segments
+          } else if (
+            matchScreen !== undefined &&
+            matchScreen < windowSegments.length
+          ) {
+            return React.cloneElement(foldableScreenChild, {
+              matchScreen: foldableScreenChild.props.matchScreen,
+            });
           }
         }
+
+        // either didn't match a custom matcher, the screen is out of range
+        // or no matching test was provided
         return null;
       })}
     </FoldableContext.Provider>