+++ CHANGELOG.md
@@ -1,18 +1,29 @@
 # Changelog for `@aaronpowell/react-foldable`
 
-## [0.2.2] - 2020-01-07
+## [0.3.0] - 2021-01-15
+
+### Changed
+
+- New release pipeline with approval workflow
+
+### Added
+
+- `render-props` option on `FoldableScreen`
+- Can pass children and function-child to `FoldableScreen`
+
+## [0.2.2] - 2021-01-07
 
 ### Fixed
 
 - Not including the output in the release package, this is now fixed
 
-## [0.2.1] - 2020-01-06
+## [0.2.1] - 2021-01-06
 
 ### Fixed
 
 - Release pipeline couldn't push to npm as a scoped package
 
-## [0.2.0] - 2020-01-06
+## [0.2.0] - 2021-01-06
 
 Released under a new package name